+++ Collection.php
@@ -1,4 +1,4 @@
-<?php namespace webmaxx\materialize;
+<?php namespace altiore\materialize;
 
 use yii\helpers\Html;
 
@@ -33,7 +33,7 @@ use yii\helpers\Html;
  * ]);
  * ```
  * @see http://materializecss.com/collections.html
- * @author webmaxx <webmaxx@webmaxx.name>
+ * @author altiore <altiore@altiore.name>
  * @since 2.0
  */
 class Collection extends Widget
