--- MaterializePluginAsset.php
@@ -1,21 +1,0 @@
-<?php namespace webmaxx\materialize;
-
-use yii\web\AssetBundle;
-
-/**
- * Asset bundle for the Materialize javascript files.
- *
- * @author webmaxx <webmaxx@webmaxx.name>
- * @since 2.0
- */
-class MaterializePluginAsset extends AssetBundle
-{
-    public $sourcePath = '@bower/materialize/dist';
-    public $js = [
-    ];
-    public $depends = [
-        'yii\web\JqueryAsset',
-        'webmaxx\materialize\MaterializeCssAsset',
-        'webmaxx\materialize\MaterializeJsAsset',
-    ];
-}
