+++ Preloader.php
@@ -1,4 +1,4 @@
-<?php namespace webmaxx\materialize;
+<?php namespace altiore\materialize;
 
 use yii\helpers\Html;
 
@@ -20,7 +20,7 @@ use yii\helpers\Html;
  * ]);
  * ```
  * @see http://materializecss.com/preloader.html
- * @author webmaxx <webmaxx@webmaxx.name>
+ * @author altiore <altiore@altiore.name>
  * @since 2.0
  */
 class Preloader extends Widget
