+++ README.md
@@ -23,13 +23,13 @@ The preferred way to install this extension is through [composer](http://getcomp
 Either run
 
 ```
-php composer.phar require --prefer-dist webmaxx/yii2-materialize
+php composer.phar require --prefer-dist altiore/yii2-materialize
 ```
 
 or add
 
 ```
-"webmaxx/yii2-materialize": "dev-master"
+"altiore/yii2-materialize": "dev-master"
 ```
 
 to the require section of your `composer.json` file.
@@ -41,5 +41,5 @@ For example, the following
 single line of code in a view file would render a Materialize Progress plugin:
 
 ```php
-<?= webmaxx\materialize\Progress::widget(['percent' => 70]) ?>
+<?= altiore\materialize\Progress::widget(['percent' => 70]) ?>
 ```
