+++ composer.json
@@ -1,27 +1,30 @@
 {
-    "name": "webmaxx/yii2-materialize",
+    "name": "altiore/yii2-materialize",
     "description": "The Materialize extension for the Yii framework",
     "keywords": ["yii2", "materialize"],
     "type": "yii2-extension",
     "license": "MIT",
     "support": {
-        "issues": "https://github.com/webmaxx/yii2-materialize/issues",
-        "source": "https://github.com/webmaxx/yii2-materialize"
+        "issues": "https://github.com/altiore/yii2-materialize/issues",
+        "source": "https://github.com/altiore/yii2-materialize"
     },
     "authors": [
+        {
+            "name": "Razzwan",
+            "email": "razvanlomov@gmail.com"
+        },
         {
             "name": "webmaxx",
             "email": "webmaxx@webmaxx.name"
         }
     ],
-    "minimum-stability": "dev",
     "require": {
-        "yiisoft/yii2": "*",
-        "bower-asset/materialize": "@stable"
+        "yiisoft/yii2": "^2.0.0",
+        "bower-asset/materialize": "^0.97.6"
     },
     "autoload": {
         "psr-4": {
-            "webmaxx\\materialize\\": ""
+            "altiore\\materialize\\": ""
         }
     },
     "extra": {
